add test to secure minimum configuration
