+++ tests/Acceptance/_minApp/.psh.yml
@@ -1,0 +1,6 @@
+paths:
+  - my/sh/scripts
+
+const: []
+
+dynamic: []
