+++ tests/Acceptance/_minApp/my/sh/scripts/test.sh
@@ -1,0 +1,3 @@
+#!/usr/bin/env bash
+
+ln -s
